requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
django-filter==2.1.0
pytz==2018.9
